refactor(ui): move SkinManager from utils/ to identity/
utils/ was a grab-bag of unrelated services (skin handling, OS file
manager invocation, game console buffer). Splitting into typed
packages by domain: identity for player-facing data (skin, future
profile-pic, future avatar overlay), platform for OS seams,
console for game-console plumbing.

SkinManager is the first move out. No behaviour change; package
declaration and three call-site imports updated.
