+++ client-ui/src/desktopMain/kotlin/hivens/ui/identity/SkinManager.kt
@@ -1,4 +1,4 @@
-package hivens.ui.utils
+package hivens.ui.identity
 
 import androidx.compose.ui.graphics.ImageBitmap
 import androidx.compose.ui.graphics.toComposeImageBitmap
