+++ client-ui/src/desktopMain/kotlin/hivens/ui/screens/ProfileScreen.kt
@@ -24,7 +24,7 @@ import hivens.ui.i18n.LocalStrings
 import hivens.ui.puppet.PuppetClick
 import hivens.ui.puppet.PuppetScreen
 import hivens.ui.theme.CelestiaTheme
-import hivens.ui.utils.SkinManager
+import hivens.ui.identity.SkinManager
 import io.github.vinceglb.filekit.FileKit
 import io.github.vinceglb.filekit.dialogs.FileKitDialogSettings
 import io.github.vinceglb.filekit.dialogs.FileKitType
